fix: bad merging
