+++ hera_cal/lst_stack/__init__.py
@@ -2,4 +2,5 @@ from .wrappers import lst_bin_files
 from .binning import lst_bin_files_for_baselines, lst_align
 from .config import make_lst_bin_config_file, make_lst_grid
 from .io import write_baseline_slc_to_file, create_lstbin_output_file
-from .averaging import sigma_clip, reduce_lst_bins, lst_average
+from .averaging import reduce_lst_bins, lst_average
+from .flagging import sigma_clip, threshold_flags
